move model classes to model paackage
