+++ kroxylicious-kms-provider-aws-kms-test-support/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/CreateAliasRequest.java
@@ -4,8 +4,9 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-record CreateKeyRequest(@JsonProperty("description") String description) {}
+public record CreateAliasRequest(@JsonProperty("TargetKeyId") String targetKeyId,
+                                 @JsonProperty("AliasName") String aliasName) {}
