+++ kroxylicious-kms-provider-aws-kms-test-support/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/CreateKeyResponse.java
@@ -4,10 +4,10 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record KeyMetadata(@JsonProperty("KeyId") String keyId, @JsonProperty("ARN") String arn) {}
+public record CreateKeyResponse(@JsonProperty("KeyMetadata") KeyMetadata keyMetadata) {}
