+++ kroxylicious-kms-provider-aws-kms-test-support/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/ScheduleKeyDeletionRequest.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+package io.kroxylicious.kms.provider.aws.kms.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record ScheduleKeyDeletionRequest(@JsonProperty(value = "KeyId") String keyId,
+                                         @JsonProperty("PendingWindowInDays") int pendingWindowInDays) {
+
+}
