+++ kroxylicious-kms-provider-aws-kms-test-support/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/ScheduleKeyDeletionResponse.java
@@ -4,12 +4,13 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-record DescribeKeyResponse(@JsonProperty(value = "KeyMetadata") KeyMetadata keyMetadata) {
+public record ScheduleKeyDeletionResponse(@JsonProperty(value = "KeyState") String keyState,
+                                          @JsonProperty(value = "PendingWindowInDays") int pendingWindowInDays) {
 
 }
