+++ kroxylicious-kms-provider-aws-kms-test-support/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/UpdateAliasRequest.java
@@ -4,11 +4,9 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-record GenerateDataKeyRequest(@JsonProperty(value = "KeyId") String keyId,
-                              @JsonProperty(value = "KeySpec") String keySpec) {
-
-}
+public record UpdateAliasRequest(@JsonProperty("TargetKeyId") String targetKeyId,
+                                 @JsonProperty("AliasName") String aliasName) {}
