--- kroxylicious-kms-provider-aws-kms/src/main/java/io/kroxylicious/kms/provider/aws/kms/DescribeKeyRequest.java
@@ -1,13 +1,0 @@
-/*
- * Copyright Kroxylicious Authors.
- *
- * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package io.kroxylicious.kms.provider.aws.kms;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-record DescribeKeyRequest(@JsonProperty(value = "KeyId") String keyId) {
-
-}
