+++ kroxylicious-kms-provider-aws-kms/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/DescribeKeyRequest.java
@@ -4,8 +4,10 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-record DeleteAliasRequest(@JsonProperty("AliasName") String aliasName) {}
+public record DescribeKeyRequest(@JsonProperty(value = "KeyId") String keyId) {
+
+}
