+++ kroxylicious-kms-provider-aws-kms/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/DescribeKeyResponse.java
@@ -4,14 +4,12 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-@SuppressWarnings("java:S6218") // we don't need DecryptResponse equality
-record DecryptResponse(@JsonProperty(value = "KeyId") String keyId,
-                       @JsonProperty(value = "Plaintext") byte[] plaintext) {
+public record DescribeKeyResponse(@JsonProperty(value = "KeyMetadata") KeyMetadata keyMetadata) {
 
 }
