+++ kroxylicious-kms-provider-aws-kms/src/main/java/io/kroxylicious/kms/provider/aws/kms/model/KeyMetadata.java
@@ -4,13 +4,10 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.kms.provider.aws.kms;
+package io.kroxylicious.kms.provider.aws.kms.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-record ScheduleKeyDeletionResponse(@JsonProperty(value = "KeyState") String keyState,
-                                   @JsonProperty(value = "PendingWindowInDays") int pendingWindowInDays) {
-
-}
+public record KeyMetadata(@JsonProperty("KeyId") String keyId, @JsonProperty("ARN") String arn) {}
